requirements.txt
Django>=5.1.6,<6.0
django-jazzmin>=3.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
qrcode[pil]>=7.4.0
Pillow>=10.0.0
drf-spectacular>=0.27.0
